apps/api/src/image-generation/requirements-diffusers.txt
diffusers>=0.35.0
transformers>=4.46.0
accelerate>=1.1.0
safetensors>=0.4.5
pillow>=10.4.0
torch>=2.1.0
